+++ lib/src/core/parser.dart
@@ -160,7 +160,7 @@ class NatrixParser {
   ) {
     final List<String> options = [];
     final List<String> args = mergedArguments;
-    ;
+    
     for (int i = 0; i < args.length; i++) {
       if (!args[i].startsWith("-")) {
         options.add(simpleStringReduction(args[i]));