+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 0.1.1
 
 - Initial version.
+
+## 0.1.2
+* Made ``NatrixStdio`` a singleton. 
+* Fixed argument parsing inside the parser and pipeline. 
+* Fixed the ``DefaultTheme`` global flags output.